feat: finishes separating dto classes
