+++ gallagher/dto/detail/items.py
@@ -1,0 +1,3 @@
+from ..utils import (
+    AppBaseModel,
+)
