+++ gallagher/dto/door.py
@@ -8,33 +8,3 @@ from .utils import (
     IdentityMixin,
     HrefMixin
 )
-
-
-class DoorRef(
-    AppBaseModel,
-    HrefMixin
-):
-    """ Door 
-
-    """
-    name: str
-
-
-class DoorSummary(
-    AppBaseModel,
-    IdentityMixin,
-    HrefMixin,
-):
-    """
-
-    """
-    name: str
-
-
-class DoorSummaryResponse(
-    AppBaseModel
-):
-    """
-
-    """
-    results: list[DoorSummary]
