+++ gallagher/dto/ref/door.py
@@ -1,0 +1,14 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin
+)
+
+
+class DoorRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ Door 
+
+    """
+    name: str
