+++ gallagher/dto/ref/event.py
@@ -1,0 +1,15 @@
+from typing import Optional
+
+from ..utils import (
+    AppBaseModel,
+    IdentityMixin,
+)
+
+
+class EventTypeSummary(
+    AppBaseModel,
+    IdentityMixin,
+):
+    """ An event type has identifiers and names
+    """
+    name: str
