+++ gallagher/dto/ref/items.py
@@ -1,0 +1,14 @@
+
+from ..utils import (
+    AppBaseModel,
+    HrefMixin
+)
+
+
+class ItemRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ Reference to an ItemType
+    """
+    name: str
