+++ gallagher/dto/ref/schedule.py
@@ -1,0 +1,17 @@
+""" Schedules
+
+"""
+
+from ..utils import (
+    AppBaseModel,
+    HrefMixin
+)
+
+
+class ScheduleRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ Schedule is a time
+    """
+    pass
