+++ gallagher/dto/ref/zone.py
@@ -1,0 +1,13 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin
+)
+
+
+class AccessZoneRef(
+    AppBaseModel,
+    HrefMixin
+):
+    """ AccessZone represents
+    """
+    name: str
