+++ gallagher/dto/response/door.py
@@ -1,0 +1,12 @@
+from ..utils import (
+    AppBaseModel,
+)
+
+
+class DoorSummaryResponse(
+    AppBaseModel
+):
+    """
+
+    """
+    results: list[DoorSummary]
