+++ gallagher/dto/response/schedule.py
@@ -1,0 +1,15 @@
+from ..utils import (
+    AppBaseModel,
+)
+
+from ..ref import (
+    ScheduleSummary
+)
+
+
+class ScheduleSummaryResponse(
+    AppBaseModel
+):
+    """ Schedule is a time
+    """
+    results: list[ScheduleSummary]
