+++ gallagher/dto/schedule.py
@@ -6,27 +6,3 @@ from .utils import (
     AppBaseModel,
     HrefMixin
 )
-
-class ScheduleRef(
-    AppBaseModel,
-    HrefMixin
-):
-    """ Schedule is a time
-    """
-    pass
-
-
-class ScheduleSummary(
-    ScheduleRef
-):
-    """ Schedule is a time
-    """
-    name: str
-
-
-class ScheduleSummaryResponse(
-    AppBaseModel
-):
-    """ Schedule is a time
-    """
-    results: list[ScheduleSummary]
