+++ gallagher/dto/summary/door.py
@@ -1,0 +1,16 @@
+from ..utils import (
+    AppBaseModel,
+    IdentityMixin,
+    HrefMixin
+)
+
+
+class DoorSummary(
+    AppBaseModel,
+    IdentityMixin,
+    HrefMixin,
+):
+    """
+
+    """
+    name: str
