+++ gallagher/dto/summary/schedule.py
@@ -1,0 +1,13 @@
+from ..utils import (
+    AppBaseModel,
+    HrefMixin
+)
+
+
+class ScheduleSummary(
+    AppBaseModel,
+    HrefMixin
+):
+    """ Schedule is a time
+    """
+    name: str
