+++ gallagher/dto/zone.py
@@ -2,17 +2,3 @@
 
 
 """
-
-from .utils import (
-    AppBaseModel,
-    HrefMixin
-)
-
-class AccessZoneRef(
-    AppBaseModel,
-    HrefMixin
-):
-    """ AccessZone represents
-    """
-    name: str
-
